app/requirements.txt
Flask==3.0.3
azure-storage-blob==12.23.1
azure-identity==1.19.0
gunicorn==23.0.0